Some additional changes Jim requested to hide PsfAttributes for DM-1067
Also removed some changes to throws which are not part of this ticket
